Move checking of conditions over to game logic
